Adding slicing to FitsImage class
This allows the user to slice an image and the slice will be applied to
all extensions of the image file. This also allows the user to perform
operations on images that do not have the same shapes (caution though
that the user must ensure that the arrays are aligned).